\name{gpl570_gpl96}
\alias{gpl570}
\alias{gpl96}

\title{matched samples on GPL570 and GPL96}

\usage{
data(gpl570_gpl96)
}

\description{
  Gene expression data of 58 matched samples on two microarrays manufactured by Affymetrix (5147 genes and 58 tumor mRNA samples) from the CellMiner, which is a web application that provides genomic profile data and query tools for the NCI60 and additional cell types. The NCI60 is a group of 60 human cancer cell lines used by the National Cancer Institute for the screening of compounds to detect potential anticancer activity. }

\value{
  \item{gpl570}{matrix of gene expression levels for the 58 matched samples measured by gpl570(Human Genome U133 Plus 2.0 Array), rows correspond to genes (5147 genes) and columns to 58 cancer cell lines.}
  \item{gpl96}{matrix of gene expression levels for the 58 matched samples measured by gpl96(Human Genome U133A array), rows correspond to genes (5147 genes) and columns to 58 cancer cell lines. }
}

\keyword{datasets}  
